apps/ops_admin/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
pywebview==4.4.1
requests==2.32.3
pyobjc==10.3.1
